Added some settings, changed stepper deenergize delay to  non-blocking++
Deenergize delay is changed to non-blocking and steppers will not be
deenergized if motion restarts before delay expires. Interrupts are no
longer reenabled in the stepper interrupt handler as in the AVR version,
this is no longer (IMO) required for ARM processors where interrupt
priority may be assigned to the different interrupt sources.